devjudge-bacekend/requirements.txt
azure-functions==1.21.3
psycopg[binary]==3.2.6
pydantic>=2,<3
